# api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: api@tinarmengineering.com
    name: Tin Arm Engineering
    url: https://www.tinarmengineering.com/
  description: The unnamed Tin Arm solver API
  license:
    name: 'Tin Arm '
  termsOfService: https://www.tinarmengineering.com/
  title: ta-solve
  version: "1.0"
servers:
- description: Chris Wallis
  url: http://localhost:3000
paths:
  /jobs/{id}:
    delete:
      description: Gets the job of the specified id
      operationId: delete-jobs-id
      parameters:
      - description: The id of the target job
        explode: false
        in: path
        name: id
        required: true
        schema: {}
        style: simple
      responses:
        "200":
          description: OK
        "404":
          description: Job not found
      summary: Delete Job
    get:
      description: Gets the job of the specified id
      operationId: get-jobs-id
      parameters:
      - description: The id of the target job
        explode: false
        in: path
        name: id
        required: true
        schema: {}
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/job'
          description: OK
        "404":
          description: Job not found
      summary: Get Job
      tags: []
    parameters:
    - description: The id of the target job
      explode: false
      in: path
      name: id
      required: true
      schema: {}
      style: simple
  /statorjobs:
    post:
      description: Creates a StatorJob with the specified properties
      operationId: post-statorjobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/stator_thermal_job'
      responses:
        "200":
          description: OK
      summary: Create StatorJob
  /rectanglejobs:
    post:
      description: Creates a RectangleJob with the specified properties
      operationId: post-rectanglejobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/rectanglejob'
      responses:
        "200":
          description: OK
      summary: Create RectangleJob
  /materials:
    post:
      operationId: post-materials
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/material'
          description: Created
      summary: Create Material
  /materials/{id}:
    delete:
      operationId: delete-materials-id
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema: {}
        style: simple
      responses:
        "200":
          description: OK
      summary: Delete Materials
    get:
      operationId: get-materials-id
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema: {}
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/material'
          description: OK
      summary: Get Materials
      tags: []
    parameters:
    - explode: false
      in: path
      name: id
      required: true
      schema: {}
      style: simple
    put:
      operationId: put-materials-id
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema: {}
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/material'
      responses:
        "200":
          description: OK
      summary: Update Materials
components:
  requestBodies: {}
  responses: {}
  schemas:
    job:
      description: Represents a basic job resource
      example:
        originalJob: ""
        id: ""
        status: ""
      properties:
        id: {}
        status: {}
        originalJob: {}
      title: job
      x-stoplight:
        id: 3w5to1qwy2bph
    stator_thermal_job:
      description: A thermal simulation of a stator tooth
      properties:
        geometry:
          allOf:
          - description: The tooth geometry
            properties:
              bore:
                description: Stator bore (mm)
              tooth_tip_depth:
                description: Tooth tip depth (mm)
              slot_opening:
                description: Slot opening (mm)
              tooth_width:
                description: Tooth width (mm)
              outer_diameter:
                description: Stator outer diameter (mm)
              back_iron_thickness:
                description: Back iron thickness (mm)
              internal_radius:
                description: Internal redius (mm)
              number_slots:
                description: Number of slots
                format: int32
              tooth_tip_angle:
                description: Tool tip angle (degrees)
            required:
            - back_iron_thickness
            - bore
            - internal_radius
            - number_slots
            - outer_diameter
            - slot_opening
            - tooth_tip_angle
            - tooth_tip_depth
            - tooth_width
            title: statorgeometry
            x-stoplight:
              id: km8rtyg43njwb
          - description: Geometric parameters of the insulation
            properties:
              slot_wedge_depth:
                default: 0.0025
                description: "The depth of the slot wedge, in m.  It can not be bigger\
                  \ than the tooth tip."
                format: double
              slot_liner_thickness:
                default: 2.0E-4
                description: The thickness of the slot liner in m.  A resonable value
                  is 0.2mm or 0.0002m.
                format: double
            required:
            - slot_liner_thickness
            - slot_wedge_depth
            title: insulatorgeometry
            x-stoplight:
              id: yync6a1bf74vp
      title: stator_thermal_job
      x-stoplight:
        id: 5yvxgcmkjtho2
    rectanglegeometry:
      description: 'The rectangle geometry for this job '
      properties:
        length:
          description: Length (mm)
        width:
          description: Width (mm)
      required:
      - length
      - width
      title: rectanglegeometry
      x-stoplight:
        id: fhtd8kb918jhw
    rectanglejob:
      description: Represents a rectangle job resource
      properties:
        geometry:
          $ref: '#/components/schemas/rectanglegeometry'
      required:
      - geometry
      title: rectanglejob
      x-stoplight:
        id: bt0s0m2ctgtd9
    quantity:
      example:
        quantity: ""
        magnitude: ""
      properties:
        magnitude:
          format: double
        quantity:
          items:
            properties:
              name:
                example: meters
              exponent:
                format: int32
            required:
            - exponent
            - name
            title: unit
            x-stoplight:
              id: h65e3t7iw16g0
      required:
      - magnitude
      title: quantity
      x-stoplight:
        id: yzdzqye0p2wc0
    material:
      description: Linear material model
      example:
        reference: ""
        electric_conductivity:
          quantity: ""
          magnitude: ""
        relative_permittivity: ""
        heat_capacity:
          quantity: ""
          magnitude: ""
        density:
          quantity: ""
          magnitude: ""
        heat_conductivity:
          quantity: ""
          magnitude: ""
        name: ""
        relative_permeability: ""
        id: ""
      properties:
        id:
          readOnly: true
        reference:
          format: uri
        heat_conductivity:
          $ref: '#/components/schemas/quantity'
        electric_conductivity:
          $ref: '#/components/schemas/quantity'
        relative_permeability:
          default: 1
        relative_permittivity:
          default: 1
        heat_capacity:
          $ref: '#/components/schemas/quantity'
        density:
          $ref: '#/components/schemas/quantity'
        name: {}
      required:
      - density
      - electric_conductivity
      - heat_capacity
      - heat_conductivity
      - name
      - relative_permeability
      - relative_permittivity
      title: material
      x-stoplight:
        id: f0co664ulll6g
  securitySchemes: {}
x-stoplight:
  id: 9id0ut8pp6oo8