# go/routers.go
/*
 * ta-solve
 *
 * The unnamed Tin Arm solver API
 *
 * API version: 1.0
 * Contact: api@tinarmengineering.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"crypto/tls"
	"encoding/json"
	"errors"
	"fmt"
	"log"
	"net/http"
	"os"

	jwtmiddleware "github.com/auth0/go-jwt-middleware"
	jwt "github.com/form3tech-oss/jwt-go"
	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
)

type Jwks struct {
	Keys []JSONWebKeys `json:"keys"`
}

type JSONWebKeys struct {
	Kty string   `json:"kty"`
	Kid string   `json:"kid"`
	Use string   `json:"use"`
	N   string   `json:"n"`
	E   string   `json:"e"`
	X5c []string `json:"x5c"`
}

// Route is the information for every URI.
type Route struct {
	// Name is the name of this Route.
	Name string
	// Method is the string for the HTTP method. ex) GET, POST etc..
	Method string
	// Pattern is the pattern of the URI.
	Pattern string
	// HandlerFunc is the handler function of this route.
	HandlerFunc gin.HandlerFunc
}

// Routes is the list of the generated Route.
type Routes []Route

var jwtMiddleWare *jwtmiddleware.JWTMiddleware

// NewRouter returns a new router.
func NewRouter() *gin.Engine {

	jwtMiddleware := jwtmiddleware.New(jwtmiddleware.Options{
		ValidationKeyGetter: func(token *jwt.Token) (interface{}, error) {
			aud := os.Getenv("AUTH0_API_AUDIENCE")
			checkAudience := token.Claims.(jwt.MapClaims).VerifyAudience(aud, false)
			if !checkAudience {
				return token, errors.New("invalid audience")
			}
			// verify iss claim
			iss := os.Getenv("AUTH0_DOMAIN")
			checkIss := token.Claims.(jwt.MapClaims).VerifyIssuer(iss, false)
			if !checkIss {
				return token, errors.New("invalid issuer")
			}

			cert, err := getPemCert(token)
			if err != nil {
				log.Fatalf("could not get cert: %+v", err)
			}

			result, _ := jwt.ParseRSAPublicKeyFromPEM([]byte(cert))
			return result, nil
		},
		SigningMethod: jwt.SigningMethodRS256,
	})

	// register our actual jwtMiddleware
	jwtMiddleWare = jwtMiddleware

	router := gin.Default()

	// Enable CORS for all origins
	config := cors.DefaultConfig()
	config.AllowAllOrigins = true
	config.AllowCredentials = true
	config.AddAllowHeaders("authorization")
	router.Use(cors.New(config))

	for _, route := range routes {
		switch route.Method {
		case http.MethodGet:
			router.GET(route.Pattern, authMiddleware(), route.HandlerFunc)
		case http.MethodPost:
			router.POST(route.Pattern, authMiddleware(), route.HandlerFunc)
		case http.MethodPut:
			router.PUT(route.Pattern, authMiddleware(), route.HandlerFunc)
		case http.MethodPatch:
			router.PATCH(route.Pattern, authMiddleware(), route.HandlerFunc)
		case http.MethodDelete:
			router.DELETE(route.Pattern, authMiddleware(), route.HandlerFunc)
		}
	}

	return router
}

// Index is the index handler.
func Index(c *gin.Context) {
	c.String(http.StatusOK, "Tin Arm | SnO2 | API root")
}

var routes = Routes{
	{
		"Index",
		http.MethodGet,
		"/",
		Index,
	},

	{
		"DeleteJobsId",
		http.MethodDelete,
		"/jobs/:id",
		DeleteJobsId,
	},

	{
		"DeleteMaterialsId",
		http.MethodDelete,
		"/materials/:id",
		DeleteMaterialsId,
	},

	{
		"GetJobsId",
		http.MethodGet,
		"/jobs/:id",
		GetJobsId,
	},

	{
		"GetMaterialsId",
		http.MethodGet,
		"/materials/:id",
		GetMaterialsId,
	},

	{
		"PostMaterials",
		http.MethodPost,
		"/materials",
		PostMaterials,
	},

	{
		"PostRectanglejobs",
		http.MethodPost,
		"/rectanglejobs",
		PostRectanglejobs,
	},

	{
		"PostStatorjobs",
		http.MethodPost,
		"/statorjobs",
		PostStatorjobs,
	},

	{
		"PutMaterialsId",
		http.MethodPut,
		"/materials/:id",
		PutMaterialsId,
	},
}

// authMiddleware intercepts the requests, and check for a valid jwt token
func authMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {

		apiKey := c.Query("apikey")

		// TODO - Check for users' own guids
		if apiKey != "" && apiKey == "testkey" {
			return
		}

		// Get the client secret key
		err := jwtMiddleWare.CheckJWT(c.Writer, c.Request)
		if err != nil {
			// Token not found
			fmt.Println(err)
			c.Abort()
			c.Writer.WriteHeader(http.StatusUnauthorized)
			c.Writer.Write([]byte("Unauthorized"))
			return
		}
	}
}

func getPemCert(token *jwt.Token) (string, error) {
	cert := ""

	// Disable security checks for this call only.  Fixes cert error (Signed by unknown authority)
	// when run from docker container.
	// TODO - Actually install the correct certificates on the container?
	tr := &http.Transport{
		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
	}
	client := &http.Client{Transport: tr}

	resp, err := client.Get(os.Getenv("AUTH0_DOMAIN") + ".well-known/jwks.json")

	if err != nil {
		return cert, err
	}
	defer resp.Body.Close()

	var jwks = Jwks{}
	err = json.NewDecoder(resp.Body).Decode(&jwks)

	if err != nil {
		return cert, err
	}

	x5c := jwks.Keys[0].X5c
	for k, v := range x5c {
		if token.Header["kid"] == jwks.Keys[k].Kid {
			cert = "-----BEGIN CERTIFICATE-----\n" + v + "\n-----END CERTIFICATE-----"
		}
	}

	if cert == "" {
		return cert, errors.New("unable to find appropriate key")
	}

	return cert, nil
}